pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = ""
authors = [
    {name = "Victor Pires",email = "victorpiressk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pytest (>=9.0.1,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
